More refactoring (split source files)
