+++ internal/forms/forms.go
@@ -7,7 +7,6 @@
 package forms
 
 import (
-	"bufio"
 	"bytes"
 	"context"
 	"encoding/json"
@@ -18,7 +17,6 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
-	"net/textproto"
 	"os"
 	"path"
 	"path/filepath"
@@ -33,7 +31,6 @@ import (
 	"github.com/la5nta/pat/internal/debug"
 	"github.com/la5nta/pat/internal/directories"
 	"github.com/la5nta/pat/internal/gpsd"
-	"github.com/la5nta/wl2k-go/fbb"
 )
 
 const formsVersionInfoURL = "https://api.getpat.io/v1/forms/standard-templates/latest"
@@ -73,17 +70,6 @@ type Config struct {
 	GPSd       cfg.GPSdConfig
 }
 
-// Template holds information about a Winlink form template
-type Template struct {
-	Name            string `json:"name"`
-	TxtFileURI      string `json:"txt_file_uri"`
-	InitialURI      string `json:"initial_uri"`
-	ViewerURI       string `json:"viewer_uri"`
-	ReplyTxtFileURI string `json:"reply_txt_file_uri"`
-	ReplyInitialURI string `json:"reply_initial_uri"`
-	ReplyViewerURI  string `json:"reply_viewer_uri"`
-}
-
 // FormFolder is a folder with forms. A tree structure with Form leaves and sub-Folder branches
 type FormFolder struct {
 	Name      string       `json:"name"`
@@ -92,33 +78,6 @@ type FormFolder struct {
 	FormCount int          `json:"form_count"`
 	Forms     []Template   `json:"forms"`
 	Folders   []FormFolder `json:"folders"`
-}
-
-// Message represents a concrete message compiled from a template
-type Message struct {
-	To      string `json:"msg_to"`
-	Cc      string `json:"msg_cc"`
-	Subject string `json:"msg_subject"`
-	Body    string `json:"msg_body"`
-
-	attachmentXML string
-	attachmentTXT string
-	targetForm    Template
-	isReply       bool
-	submitted     time.Time
-}
-
-// Attachments returns the attachments generated by the filled-in form
-func (m Message) Attachments() []*fbb.File {
-	var files []*fbb.File
-	if xml := m.attachmentXML; xml != "" {
-		name := getXMLAttachmentNameForForm(m.targetForm, m.isReply)
-		files = append(files, fbb.NewFile(name, []byte(xml)))
-	}
-	if txt := m.attachmentTXT; txt != "" {
-		files = append(files, fbb.NewFile("FormData.txt", []byte(txt)))
-	}
-	return files
 }
 
 // UpdateResponse is the API response format for the upgrade forms endpoint
@@ -348,20 +307,6 @@ func (m *Manager) downloadAndUnzipForms(ctx context.Context, downloadLink string
 	return nil
 }
 
-// getXMLAttachmentNameForForm returns the user-visible filename for the message attachment that holds the form instance values
-func getXMLAttachmentNameForForm(t Template, isReply bool) string {
-	attachmentName := filepath.Base(t.ViewerURI)
-	if isReply {
-		attachmentName = filepath.Base(t.ReplyViewerURI)
-	}
-	attachmentName = strings.TrimSuffix(attachmentName, filepath.Ext(attachmentName))
-	attachmentName = "RMS_Express_Form_" + attachmentName + ".xml"
-	if len(attachmentName) > 255 {
-		attachmentName = strings.TrimPrefix(attachmentName, "RMS_Express_Form_")
-	}
-	return attachmentName
-}
-
 // RenderForm finds the associated form and returns the filled-in form in HTML given the contents of a form attachment
 func (m *Manager) RenderForm(data []byte, composeReply bool) (string, error) {
 	type Node struct {
@@ -461,28 +406,6 @@ func (m *Manager) ComposeTemplate(tmplPath string, subject string) (Message, err
 	}.build()
 }
 
-func (t Template) matchesName(nameToMatch string) bool {
-	return t.InitialURI == nameToMatch ||
-		strings.EqualFold(t.InitialURI, nameToMatch+htmlFileExt) ||
-		t.ViewerURI == nameToMatch ||
-		strings.EqualFold(t.ViewerURI, nameToMatch+htmlFileExt) ||
-		t.ReplyInitialURI == nameToMatch ||
-		t.ReplyInitialURI == nameToMatch+".0" ||
-		t.ReplyViewerURI == nameToMatch ||
-		t.ReplyViewerURI == nameToMatch+".0" ||
-		t.TxtFileURI == nameToMatch ||
-		strings.EqualFold(t.TxtFileURI, nameToMatch+txtFileExt)
-}
-
-func (t Template) containsName(partialName string) bool {
-	return strings.Contains(t.InitialURI, partialName) ||
-		strings.Contains(t.ViewerURI, partialName) ||
-		strings.Contains(t.ReplyInitialURI, partialName) ||
-		strings.Contains(t.ReplyViewerURI, partialName) ||
-		strings.Contains(t.ReplyTxtFileURI, partialName) ||
-		strings.Contains(t.TxtFileURI, partialName)
-}
-
 func (m *Manager) buildFormFolder() (FormFolder, error) {
 	formFolder, err := m.innerRecursiveBuildFormFolder(m.config.FormsPath)
 	formFolder.Version = m.getFormsVersion()
@@ -555,93 +478,6 @@ func (m *Manager) rel(path string) string {
 	return rel
 }
 
-// resolveFileReference searches for files referenced in .txt files.
-//
-// If found the returned path is relative to FormsPath and bool is true, otherwise the given path is returned unmodified.
-func resolveFileReference(basePath string, referencePath string) (string, bool) {
-	path := filepath.Join(basePath, referencePath)
-	if !directories.IsInPath(basePath, path) {
-		debug.Printf("%q escapes template's base path (%q)", referencePath, basePath)
-		return "", false
-	}
-	if _, err := os.Stat(path); err == nil {
-		return path, true
-	}
-	// Fallback to case-insenstive search.
-	// Some HTML files references in the .txt files has a different caseness than the actual filename on disk.
-	//TODO: Walk basePath tree instead
-	absPathTemplateFolder := filepath.Dir(path)
-	entries, err := os.ReadDir(absPathTemplateFolder)
-	if err != nil {
-		return path, false
-	}
-	for _, entry := range entries {
-		name := entry.Name()
-		if strings.EqualFold(filepath.Base(path), name) {
-			return filepath.Join(absPathTemplateFolder, name), true
-		}
-	}
-	return path, false
-}
-
-func readTemplate(path string) (Template, error) {
-	f, err := os.Open(path)
-	if err != nil {
-		return Template{}, err
-	}
-	defer f.Close()
-
-	template := Template{
-		Name:       strings.TrimSuffix(filepath.Base(path), filepath.Ext(path)),
-		TxtFileURI: path,
-	}
-	scanner := bufio.NewScanner(f)
-	baseURI := filepath.Dir(template.TxtFileURI)
-	for scanner.Scan() {
-		switch key, value, _ := strings.Cut(scanner.Text(), ":"); key {
-		case "Form":
-			// Form: <composer>,<viewer>
-			files := strings.Split(value, ",")
-			// Extend to absolute paths and add missing html extension
-			for i, path := range files {
-				path = strings.TrimSpace(path)
-				if ext := filepath.Ext(path); ext == "" {
-					path += htmlFileExt
-				}
-				var ok bool
-				files[i], ok = resolveFileReference(baseURI, path)
-				if !ok {
-					debug.Printf("%s: failed to resolve referenced file %q", template.TxtFileURI, path)
-				}
-			}
-			template.InitialURI = files[0]
-			if len(files) > 1 {
-				template.ViewerURI = files[1]
-			}
-		case "ReplyTemplate":
-			path := strings.TrimSpace(value)
-			// Some are missing .txt
-			if filepath.Ext(path) == "" {
-				path += txtFileExt
-			}
-			var ok bool
-			path, ok = resolveFileReference(baseURI, path)
-			if !ok {
-				debug.Printf("%s: failed to resolve referenced reply template file %q", template.TxtFileURI, path)
-				continue
-			}
-			replyTemplate, err := readTemplate(path)
-			if err != nil {
-				debug.Printf("%s: failed to load referenced reply template: %v", template.TxtFileURI, err)
-			}
-			template.ReplyTxtFileURI = path
-			template.ReplyInitialURI = replyTemplate.InitialURI
-			template.ReplyViewerURI = replyTemplate.ViewerURI
-		}
-	}
-	return template, err
-}
-
 func findFormFromURI(formName string, folder FormFolder) (Template, error) {
 	form := Template{Name: "unknown"}
 	for _, subFolder := range folder.Folders {
@@ -701,7 +537,7 @@ func (m *Manager) fillFormTemplate(tmplPath string, formDestURL string, formVars
 	data = strings.ReplaceAll(data, "http://localhost:8001", formDestURL) // Some Canada BC forms are hardcoded to this URL
 
 	// Substitute insertion tags and variables
-	data = m.insertionTagReplacer("{", "}")(data)
+	data = insertionTagReplacer(m, "{", "}")(data)
 	data = variableReplacer("{", "}", formVars)(data)
 
 	return data, nil
@@ -714,268 +550,6 @@ func (m *Manager) getFormsVersion() string {
 		return "unknown"
 	}
 	return strings.TrimSpace(str)
-}
-
-type messageBuilder struct {
-	Interactive bool
-	IsReply     bool
-	Template    Template
-	FormValues  map[string]string
-	FormsMgr    *Manager
-}
-
-// build returns message subject, body, and attachments for the given template and variable map
-func (b messageBuilder) build() (Message, error) {
-	tmplPath := b.Template.TxtFileURI
-	if b.IsReply && b.Template.ReplyTxtFileURI != "" {
-		tmplPath = b.Template.ReplyTxtFileURI
-	}
-
-	b.initFormValues()
-
-	msgForm, err := b.scanTmplBuildMessage(tmplPath)
-	if err != nil {
-		return Message{}, err
-	}
-	//TODO: Should any of these be set in scanTmplBuildMessage?
-	msgForm.targetForm = b.Template
-	msgForm.isReply = b.IsReply
-	msgForm.submitted = time.Now()
-
-	// Add TXT attachment if defined by the form
-	if v, ok := b.FormValues["attached_text"]; ok {
-		delete(b.FormValues, "attached_text")
-		msgForm.attachmentTXT = v
-	}
-
-	formVarsAsXML := ""
-	for varKey, varVal := range b.FormValues {
-		formVarsAsXML += fmt.Sprintf("    <%s>%s</%s>\n", xmlEscape(varKey), xmlEscape(varVal), xmlEscape(varKey))
-	}
-
-	// Add XML if a viewer is defined for this form
-	if b.Template.ViewerURI != "" {
-		viewer := b.Template.ViewerURI
-		if b.IsReply && b.Template.ReplyViewerURI != "" {
-			viewer = b.Template.ReplyViewerURI
-		}
-		msgForm.attachmentXML = fmt.Sprintf(`%s<RMS_Express_Form>
-  <form_parameters>
-    <xml_file_version>%s</xml_file_version>
-    <rms_express_version>%s</rms_express_version>
-    <submission_datetime>%s</submission_datetime>
-    <senders_callsign>%s</senders_callsign>
-    <grid_square>%s</grid_square>
-    <display_form>%s</display_form>
-    <reply_template>%s</reply_template>
-  </form_parameters>
-  <variables>
-%s
-  </variables>
-</RMS_Express_Form>
-`,
-			xml.Header,
-			"1.0",
-			b.FormsMgr.config.AppVersion,
-			time.Now().UTC().Format("20060102150405"),
-			b.FormsMgr.config.MyCall,
-			b.FormsMgr.config.Locator,
-			filepath.Base(viewer),
-			filepath.Base(b.Template.ReplyTxtFileURI),
-			formVarsAsXML)
-	}
-
-	msgForm.To = strings.TrimSpace(msgForm.To)
-	msgForm.Cc = strings.TrimSpace(msgForm.Cc)
-	msgForm.Subject = strings.TrimSpace(msgForm.Subject)
-	msgForm.Body = strings.TrimSpace(msgForm.Body)
-	return msgForm, nil
-}
-
-func (b messageBuilder) initFormValues() {
-	if b.IsReply {
-		b.FormValues["msgisreply"] = "True"
-	} else {
-		b.FormValues["msgisreply"] = "False"
-	}
-	for _, key := range []string{"msgsender"} {
-		if _, ok := b.FormValues[key]; !ok {
-			b.FormValues[key] = b.FormsMgr.config.MyCall
-		}
-	}
-
-	// some defaults that we can't set yet. Winlink doesn't seem to care about these
-	// Set only if they're not set by form values.
-	for _, key := range []string{"msgto", "msgcc", "msgsubject", "msgbody", "msgp2p", "txtstr"} {
-		if _, ok := b.FormValues[key]; !ok {
-			b.FormValues[key] = ""
-		}
-	}
-	for _, key := range []string{"msgisforward", "msgisacknowledgement"} {
-		if _, ok := b.FormValues[key]; !ok {
-			b.FormValues[key] = "False"
-		}
-	}
-
-	//TODO: Implement sequences
-	for _, key := range []string{"msgseqnum"} {
-		if _, ok := b.FormValues[key]; !ok {
-			b.FormValues[key] = "0"
-		}
-	}
-}
-
-func (b messageBuilder) scanTmplBuildMessage(tmplPath string) (Message, error) {
-	f, err := os.Open(tmplPath)
-	if err != nil {
-		return Message{}, err
-	}
-	defer f.Close()
-
-	replaceVars := variableReplacer("<", ">", b.FormValues)
-	replaceInsertionTags := b.FormsMgr.insertionTagReplacer("<", ">")
-
-	scanner := bufio.NewScanner(f)
-
-	var msgForm Message
-	var inBody bool
-	for scanner.Scan() {
-		lineTmpl := scanner.Text()
-
-		// Insertion tags and variables
-		lineTmpl = replaceInsertionTags(lineTmpl)
-		lineTmpl = replaceVars(lineTmpl)
-
-		// Prompts (mostly found in text templates)
-		if b.Interactive {
-			lineTmpl = promptAsks(lineTmpl, func(a Ask) string {
-				//TODO: Handle a.Multiline as we do message body
-				fmt.Printf(a.Prompt + " ")
-				return b.FormsMgr.config.LineReader()
-			})
-			lineTmpl = promptSelects(lineTmpl, func(s Select) Option {
-				for {
-					fmt.Println(s.Prompt)
-					for i, opt := range s.Options {
-						fmt.Printf("  %d\t%s\n", i, opt.Item)
-					}
-					fmt.Printf("select 0-%d: ", len(s.Options)-1)
-					idx, err := strconv.Atoi(b.FormsMgr.config.LineReader())
-					if err == nil && idx < len(s.Options) {
-						return s.Options[idx]
-					}
-				}
-			})
-			// Fallback prompt for undefined form variables.
-			// Typically these are defined by the associated HTML form, but since
-			// this is CLI land we'll just prompt for the variable value.
-			lineTmpl = promptVars(lineTmpl, func(key string) string {
-				fmt.Println(lineTmpl)
-				fmt.Printf("%s: ", key)
-				value := b.FormsMgr.config.LineReader()
-				b.FormValues[strings.ToLower(key)] = value
-				replaceVars = variableReplacer("<", ">", b.FormValues) // Refresh variableReplacer to avoid prompting the same again
-				return value
-			})
-		}
-
-		if inBody {
-			msgForm.Body += lineTmpl + "\n"
-			continue // No control fields in body
-		}
-
-		// Control fields
-		switch key, value, _ := strings.Cut(lineTmpl, ":"); textproto.CanonicalMIMEHeaderKey(key) {
-		case "Msg":
-			// The message body starts here. No more control fields after this.
-			msgForm.Body += value
-			inBody = true
-		case "Form", "ReplyTemplate":
-			// Handled elsewhere
-			continue
-		case "Def", "Define":
-			// Def: variable=value – Define the value of a variable.
-			key, value, ok := strings.Cut(value, "=")
-			if !ok {
-				debug.Printf("Def: without key-value pair: %q", value)
-				continue
-			}
-			key, value = strings.ToLower(strings.TrimSpace(key)), strings.TrimSpace(value)
-			b.FormValues[key] = value
-			debug.Printf("Defined %q=%q", key, value)
-		case "Subject", "Subj":
-			// Set the subject of the message
-			msgForm.Subject = strings.TrimSpace(value)
-		case "To":
-			// Specify to whom the message is being sent
-			msgForm.To = strings.TrimSpace(value)
-		case "Cc":
-			// Specify carbon copy addresses
-			msgForm.Cc = strings.TrimSpace(value)
-		case "Readonly":
-			// Yes/No – Specify whether user can edit.
-			// TODO: Disable editing of body in composer?
-		case "Seqinc":
-			//TODO: Handle sequences
-		default:
-			if strings.TrimSpace(lineTmpl) != "" {
-				log.Printf("skipping unknown template line: '%s'", lineTmpl)
-			}
-		}
-	}
-	return msgForm, nil
-}
-
-// VariableReplacer returns a function that replaces the given key-value pairs.
-func variableReplacer(tagStart, tagEnd string, vars map[string]string) func(string) string {
-	return placeholderReplacer(tagStart+"Var ", tagEnd, vars)
-}
-
-// InsertionTagReplacer returns a function that replaces the fixed set of insertion tags with their corresponding values.
-func (m *Manager) insertionTagReplacer(tagStart, tagEnd string) func(string) string {
-	now := time.Now()
-	validPos := "NO"
-	nowPos, err := m.gpsPos()
-	if err != nil {
-		debug.Printf("GPSd error: %v", err)
-	} else {
-		validPos = "YES"
-		debug.Printf("GPSd position: %s", positionFmt(signedDecimal, nowPos))
-	}
-	return placeholderReplacer(tagStart, tagEnd, map[string]string{
-		"MsgSender":      m.config.MyCall,
-		"Callsign":       m.config.MyCall,
-		"ProgramVersion": "Pat " + m.config.AppVersion,
-
-		"DateTime":  formatDateTime(now),
-		"UDateTime": formatDateTimeUTC(now),
-		"Date":      formatDate(now),
-		"UDate":     formatDateUTC(now),
-		"UDTG":      formatUDTG(now),
-		"Time":      formatTime(now),
-		"UTime":     formatTimeUTC(now),
-
-		"GPS":                positionFmt(degreeMinute, nowPos),
-		"GPS_DECIMAL":        positionFmt(decimal, nowPos),
-		"GPS_SIGNED_DECIMAL": positionFmt(signedDecimal, nowPos),
-		"Latitude":           fmt.Sprintf("%.4f", nowPos.Lat),
-		"Longitude":          fmt.Sprintf("%.4f", nowPos.Lon),
-		"GridSquare":         posToGridSquare(nowPos),
-		"GPSValid":           fmt.Sprintf("%s ", validPos),
-
-		//TODO (other insertion tags found in Standard Forms):
-		// SeqNum
-		// FormFolder
-		// GPSLatitude
-		// GPSLongitude
-		// InternetAvailable
-		// MsgP2P
-		// MsgSubject
-		// Sender
-		// Speed
-		// course
-		// decimal_separator
-	})
 }
 
 func xmlEscape(s string) string {
